requirements-gpu.txt
# GPU acceleration requirements for text-extract-api
# Install these for GPU-accelerated text extraction

# Core GPU support packages
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Install PyTorch with CUDA support (uncomment for NVIDIA GPU):
# --extra-index-url https://download.pytorch.org/whl/cu118
# torch>=2.0.0+cu118
# torchvision>=0.15.0+cu118
# torchaudio>=2.0.0+cu118

# TensorFlow with GPU support
tensorflow[and-cuda]>=2.12.0

# OpenCL support (for AMD/Intel GPUs)
pyopencl>=2023.1

# Enhanced image processing
pillow-simd>=9.0.0  # Faster PIL operations
opencv-python-headless>=4.7.0

# CUDA Python (for direct CUDA operations)
cuda-python>=12.0.0

# GPU monitoring
nvidia-ml-py3>=7.352.0  # For NVIDIA GPU monitoring
gpustat>=1.1.0  # GPU usage statistics

# Memory management
memory-profiler>=0.60.0

# Optional: Jupyter with GPU support
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Note: EasyOCR will automatically use GPU when available
# No additional packages needed for EasyOCR GPU support
